+++ pkg/integrations/install/install.go
@@ -19,6 +19,7 @@ import (
 	_ "github.com/grafana/agent/pkg/integrations/mongodb_exporter"       // register mongodb_exporter
 	_ "github.com/grafana/agent/pkg/integrations/mysqld_exporter"        // register mysqld_exporter
 	_ "github.com/grafana/agent/pkg/integrations/node_exporter"          // register node_exporter
+	_ "github.com/grafana/agent/pkg/integrations/oracledb_exporter"      // register oracledb_exporter
 	_ "github.com/grafana/agent/pkg/integrations/postgres_exporter"      // register postgres_exporter
 	_ "github.com/grafana/agent/pkg/integrations/process_exporter"       // register process_exporter
 	_ "github.com/grafana/agent/pkg/integrations/redis_exporter"         // register redis_exporter
@@ -36,6 +37,7 @@ import (
 	_ "github.com/grafana/agent/pkg/integrations/v2/app_agent_receiver" // register app_agent_receiver
 	_ "github.com/grafana/agent/pkg/integrations/v2/blackbox_exporter"  // register blackbox_exporter
 	_ "github.com/grafana/agent/pkg/integrations/v2/eventhandler"       // register eventhandler
+	_ "github.com/grafana/agent/pkg/integrations/v2/oracledb_exporter"  // register oracledb_exporter
 	_ "github.com/grafana/agent/pkg/integrations/v2/snmp_exporter"      // register snmp_exporter
 	_ "github.com/grafana/agent/pkg/integrations/v2/snowflake"          // register snowflake
 	_ "github.com/grafana/agent/pkg/integrations/v2/vmware_exporter"    // register vmware_exporter