+++ pkg/integrations/v2/oracledb_exporter/oracledb_exporter_test.go
@@ -1,0 +1,6 @@
+package oracledbexporter
+
+import "testing"
+
+func TestOracleDBConfig(t *testing.T) {
+}